move parse_artists_runs out of _utils.py
